ADD: Empleado View
Refactor: Empleado Entity
ADD: Empleado view
UDP: Empleado Frame

+++ src/Pack_negocio/EmpleadoJpaController.java
@@ -9,6 +9,7 @@ import Pack_negocio.exceptions.NonexistentEntityException;
 import Pack_negocio.exceptions.PreexistingEntityException;
 import Pack_persistencia.Empleado;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
@@ -61,7 +62,7 @@ public class EmpleadoJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Long id = empleado.getIdEmpleado();
+                BigDecimal id = empleado.getIdEmpleado();
                 if (findEmpleado(id) == null) {
                     throw new NonexistentEntityException("The empleado with id " + id + " no longer exists.");
                 }
@@ -74,7 +75,7 @@ public class EmpleadoJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(BigDecimal id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -119,7 +120,7 @@ public class EmpleadoJpaController implements Serializable {
         }
     }
 
-    public Empleado findEmpleado(Long id) {
+    public Empleado findEmpleado(BigDecimal id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Empleado.class, id);
